Modularize program_turnon.py
